requirements.txt
python-decouple
fastapi
uvicorn
requests
python-multipart